+++ packages/genui/lib/src/core/a2ui_message_processor.dart
@@ -13,9 +13,9 @@ import '../model/chat_message.dart';
 import '../model/data_model.dart';
 import '../model/ui_models.dart';
 import '../primitives/logging.dart';
-import 'genui_configuration.dart';
 
-/// A sealed class representing an update to the UI managed by [GenUiManager].
+/// A sealed class representing an update to the UI managed by
+/// [A2uiMessageProcessor].
 ///
 /// This class has three subclasses: [SurfaceAdded], [SurfaceUpdated], and
 /// [SurfaceRemoved].
@@ -84,14 +84,10 @@ abstract interface class GenUiHost {
 /// `UiDefinition`. It provides the tools (`surfaceUpdate`, `deleteSurface`,
 /// `beginRendering`) that the AI uses to manipulate the UI. It exposes a stream
 /// of `GenUiUpdate` events so that the application can react to changes.
-class GenUiManager implements GenUiHost {
-  /// Creates a new [GenUiManager] with a list of supported widget catalogs.
-  GenUiManager({
-    required this.catalogs,
-    this.configuration = const GenUiConfiguration(),
-  });
-
-  final GenUiConfiguration configuration;
+class A2uiMessageProcessor implements GenUiHost {
+  /// Creates a new [A2uiMessageProcessor] with a list of supported widget
+  /// catalogs.
+  A2uiMessageProcessor({required this.catalogs});
 
   @override
   final Iterable<Catalog> catalogs;